refactor: JWT 및 쿠키 유효 시간 타입 수정 (#71)
- JWT 유효 시간 필드 타입을 Integer에서 Long으로 변경
- JwtUtil, CookieUtil, RefreshTokenService에서 관련 메서드 및 필드 수정
- AuthenticationResponse의 쿠키 유효 시간 필드 타입 수정으로 일관성 유지

+++ src/main/java/com/example/moabackend/global/security/info/AuthenticationResponse.java
@@ -19,14 +19,9 @@ public class AuthenticationResponse {
             HttpServletResponse response,
             String domain,
             JwtDTO jwtDto,
-            Integer refreshExpiration
+            Long refreshExpiration
     ) throws IOException {
-//        CookieUtil.addCookie(
-//                response,
-//                domain,
-//                Constants.ACCESS_COOKIE_NAME,
-//                jwtDto.accessToken()
-//        );
+
         CookieUtil.addSecureCookie(
                 response,
                 domain,
